Matrix composition now works (check tests/operators for a laplacian) and made the code more gcc-compliant (not that the build will definitely build under gcc4.7, but it did work half way through this commit).  The next commits should be to clean up the code + produce some simple examples + perhaps some tools for things like particle advection (kd-trees + interpolation within simplices)

+++ include/simplicialComplex.hpp
@@ -162,7 +162,7 @@ protected:
             int M = EmbeddedDim;
             typename NumTraits::DynamicMatrix m(M,clist.size()-N-1);
             auto&& origin = clist.back();
-            for(int i=N; i < clist.size()-1; ++i)
+            for(int i=N; i < static_cast<int>(clist.size())-1; ++i)
             {
                 m.col(i-N) = this->m_vertices[i] - origin;
             }
@@ -283,7 +283,7 @@ protected:
         int M = EmbeddedDim;
         typename NumTraits::DynamicMatrix m(M,clist.size()-1);
         auto&& origin = clist.back();
-        for(int i=0; i < clist.size()-1; ++i)
+        for(int i=0; i < static_cast<int>(clist.size())-1; ++i)
         {
             m.col(i) = this->m_vertices[i] - origin;
         }
